Moderate refactor needed to create an installable package
