+++ README.md
@@ -1,12 +1,12 @@
 # ggr-cwl-ipynb-gen
 Jupyter notebook generator to download and execute the processing files for GGR related datasets. 
 At this point, is not intented to cover all use cases, but to serve as a quick generator of all 
-related files and scripts to pre-process sequences generated at the [Duke-GCB Sequencing Core](https://genome.duke.edu/cores-and-services/sequencing-and-genomic-technologies) in [HARDAC](https://genome.duke.edu/cores-and-services/computational-solutions/compute-environments-genomics).
+related files and scripts to pre-process genomic data generated at the [Duke-GCB Sequencing Core](https://genome.duke.edu/cores-and-services/sequencing-and-genomic-technologies) in [HARDAC](https://genome.duke.edu/cores-and-services/computational-solutions/compute-environments-genomics).
 
 Example of usage:
 ```
 $ python ggr-cwl-ipynb-gen.py \
-  --conf examples/conf.yaml \
+  --root-dir /path/to/rootdir \
   --metadata examples/Hong_3979_170316B1.xlsx \
   --out /path/to/output_dir \
   --force
@@ -16,33 +16,26 @@ The information in the example metadata and configuration file should reveal wha
 For a full list of options:
 ```
 $ python ggr-cwl-ipynb-gen.py -h
-usage: Generator of Jupyter notebooks to execute CWL pre-processing pipelines
-       [-h] -o OUT -c CONF_FILE -m METADATA [-f] [-n] [--metadata-sep SEP]
-       [--project-name PROJECT_NAME] [--data-from {sftp,miseq,other,dukeds}]
+usage: Generator of Jupyter notebooks to execute CWL pre-processing pipelines [-h] -o OUT -m METADATA [-f] [-n] [--metadata-sep SEP] [--project-name PROJECT_NAME]
+                                                                              [--data-from {sftp,miseq,other,dukeds,local}] [-c CONF_FILE] [-u USER] [-e USER_DUKE_EMAIL] [-r ROOT_DIR]
 
 optional arguments:
   -h, --help            show this help message and exit
   -o OUT, --out OUT     Jupyter notebook output file name
-  -c CONF_FILE, --conf-file CONF_FILE
-                        YAML configuration file (see examples)
   -m METADATA, --metadata METADATA
                         Metadata file with samples information
   -f, --force           Force to overwrite output file
-  -n, --no-upload       Avoids uploading generated data to database when
-                        specified
-  --metadata-sep SEP    Separator for metadata file (when different than Excel
-                        spread sheet)
+  -n, --no-upload       Avoids uploading generated data to database when specified
+  --metadata-sep SEP    Separator for metadata file (when different than Excel spread sheet)
   --project-name PROJECT_NAME
-                        Project name (by default, basename of metadata file
-                        name)
-  --data-from {sftp,miseq,other,dukeds}
-                        Choices: sftp, miseq, other, dukeds
+                        Project name (by default, basename of metadata file name)
+  --data-from {sftp,miseq,other,dukeds,local}
+                        Choices: sftp, miseq, other, dukeds, local
+  -c CONF_FILE, --conf-file CONF_FILE
+                        YAML configuration file (see examples)
+  -u USER, --user USER  HARDAC User used in SLURM (default: ${USER})
+  -e USER_DUKE_EMAIL, --user-duke-email USER_DUKE_EMAIL
+                        Email(s) notified when execution is finished (default: ${USER}@duke.edu)
+  -r ROOT_DIR, --root-dir ROOT_DIR
+                        Root directory where all subfolders and files will be created (semi-required: either defined here or in conf-file)
 ```
-
-### Dependencies
-- jinja2 >=2.8
-- nbformat >=4.0.1
-- numpy >=1.10.4
-- pandas >=0.17.1
-- xlrd >=1.0.0
-- ruamel >=0.11.11
