+++ VERSION.py
@@ -1,0 +1,1 @@
+VERSION = '0.5.0'
