+++ ggr_cwl_ipynb_gen/__init__.py
@@ -1,0 +1,2 @@
+if __name__ == '__main__':
+    ggr-main()
