+++ ggr_cwl_ipynb_gen/consts.py
@@ -53,3 +53,6 @@ plot_script = '/data/reddylab/Darryl/GitHub/reddylab/countFactors_metadata.sh'
 qc_script_dir = '/data/reddylab/software/cwl/bin'
 data_upload_script = '/data/reddylab/Darryl/GitHub/reddylab/csv_to_mongo.py'
 HOST_FOR_TUNNELED_DOWNLOAD = "Hardac-xfer.genome.duke.edu"
+
+# Package constants
+PACKAGE_NAME = "ggr_cwl_ipynb_gen"
