+++ ggr_cwl_ipynb_gen/templates/cwl_json_gen.j2
