+++ ggr_cwl_ipynb_gen/templates/cwl_slurm_array_gen.j2
