+++ ggr_cwl_ipynb_gen/templates/data_upload.j2
