+++ ggr_cwl_ipynb_gen/templates/download_fastq_files.j2
