+++ ggr_cwl_ipynb_gen/templates/generate_plots.j2
