+++ ggr_cwl_ipynb_gen/templates/generate_qc_cell.j2
