+++ ggr_cwl_ipynb_gen/templates/ungzip_fastq_files.j2
