+++ requirements.txt
@@ -1,0 +1,9 @@
+jinja2 >=2.8
+nbformat >=4.0.1
+numpy >=1.10.4
+pandas >=0.17.1
+xlrd >=1.0.0
+ruamel.yaml >=0.11.11
+
+setuptools
+pymongo
