+++ setup.cfg
@@ -1,0 +1,18 @@
+[metadata]
+name = ggr-cwl-ipynb-gen-alexbarrera
+version = 0.5.0
+author = Alejandro Barrera
+author_email = alejandro.barrera@duke.edu
+description = IPython notebook generator for GGR CWL processing pipelines of genomic data
+long_description = file: README.md
+long_description_content_type = text/markdown
+url = https://github.com/alexbarrera/ggr-cwl-ipynb-gen
+project_urls =
+    Bug Tracker = https://github.com/alexbarrera/ggr-cwl-ipynb-gen/issues
+classifiers =
+    Programming Language :: Python :: 3
+    License :: OSI Approved :: MIT License
+    Operating System :: OS Independent
+
+[options]
+packages = find:
